# Conflicts:
#	Assets/GameSparks/Plugins/NX64/GameSparksNative.a
#	Assets/Resources/HUFConfigs/AdMobProviderConfig.asset
#	remote_config_data

+++ Assets/HUF/Utils/Editor/AssetsBuilder/HUFBuildAssetsResolver.cs
@@ -1,10 +1,13 @@
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using HUF.Utils.Runtime.Extensions;
 using HUF.Utils.Runtime.Logging;
 using UnityEditor;
 using UnityEditor.Build;
 using UnityEditor.Build.Reporting;
+using UnityEditor.Callbacks;
 using UnityEngine;
 
 namespace HUF.Utils.Editor.AssetsBuilder
@@ -15,14 +18,15 @@ namespace HUF.Utils.Editor.AssetsBuilder
         const string ANDROID_PLUGINS_FOLDER = PLUGIN_FOLDER + "Android/";
         const string IOS_PLUGINS_FOLDER = PLUGIN_FOLDER + "iOS/";
 
-        const string HUF_ANDROID_ASSET_FOLDER = "Assets/Android";
-        const string HUF_IOS_ASSET_FOLDER = "Assets/iOS/";
+        const string HUF_ANDROID_ASSET_FOLDER = "Plugins/Android";
+        const string HUF_IOS_ASSET_FOLDER = "Plugins/iOS";
         const string MANIFEST_COPY_SEARCH = "Manifest.huf";
         public const string MANIFEST_TEMPLATE_SEARCH = "ManifestTemplate.huf";
 
         public const string MANIFEST_FULL_NAME = "AndroidManifest.xml";
         const string PROJECT_PROPERTY_FULL_NAME = "project.properties";
         const string ANDROID_ADDITIONAL_MANIFEST_FOLDER_END_NAME = ".androidlib";
+        const string META_FILES = ".meta";
 
         static readonly HLogPrefix logPrefix = new HLogPrefix( nameof(HUFBuildAssetsResolver) );
 
@@ -32,8 +36,14 @@ namespace HUF.Utils.Editor.AssetsBuilder
         {
             RunHUFBuildAssetsResolver();
         }
-        
-        [MenuItem( "HUF/Tools/Build Assets Resolver" )]
+
+        [PostProcessBuildAttribute( 1000 )]
+        public static void PostProcessBuildAttribute( BuildTarget target, string pathToBuiltProject )
+        {
+            RunHUFBuildAssetsReverter();
+        }
+
+        [MenuItem( "HUF/Utils/Builds/Build Assets Resolver" )]
         public static void RunHUFBuildAssetsResolver()
         {
 #if UNITY_ANDROID
@@ -41,15 +51,76 @@ namespace HUF.Utils.Editor.AssetsBuilder
 #endif
         }
 
+        [MenuItem( "HUF/Utils/Builds/Build Assets Revert" )]
+        public static void RunHUFBuildAssetsReverter()
+        {
+#if UNITY_ANDROID
+            HUFAndroidBuildsAssetsReverter();
+#endif
+        }
+
+        static void HUFAndroidBuildsAssetsReverter()
+        {
+            var paths = GetHUFManifests( MANIFEST_COPY_SEARCH );
+
+            foreach ( string sourcePath in paths )
+            {
+                var pathSplit = sourcePath.Split( '/' );
+                //var packageAndroidFolder = sourcePath.Replace( pathSplit[pathSplit.Length - 1], "" );
+                var packageEndPath = GetAndroidManifestEndPath( sourcePath );
+                FileUtil.DeleteFileOrDirectory( packageEndPath );
+                FileUtil.DeleteFileOrDirectory( packageEndPath + META_FILES );
+            }
+
+            paths = GetHUFManifests( MANIFEST_TEMPLATE_SEARCH );
+
+            foreach ( string sourcePath in paths )
+            {
+                var pathSplit = sourcePath.Split( '/' );
+                var packageEndPath = GetAndroidManifestEndPath( sourcePath );
+                FileUtil.DeleteFileOrDirectory( packageEndPath );
+                FileUtil.DeleteFileOrDirectory( packageEndPath + META_FILES );
+            }
+
+            paths = GetHUFPackageToCopyFolder();
+
+            foreach ( string source in paths )
+            {
+                var sourcePath = source + "ToCopy/";
+
+                if ( !Directory.Exists( sourcePath ) )
+                    continue;
+
+                DeleteAllDirectorys( sourcePath );
+            }
+            
+            if ( Directory.Exists( $"{ANDROID_PLUGINS_FOLDER}Firebase" ) )
+                DeleteAllDirectorys( $"{ANDROID_PLUGINS_FOLDER}Firebase" , false);
+        }
+
+        static void DeleteAllDirectorys( string sourcePath, bool addAndroidLibraryName = true )
+        {
+            var dir = new DirectoryInfo( sourcePath );
+            var directories = dir.GetDirectories();
+
+            foreach ( var directory in directories )
+            {
+                if ( directory.Name == "res" )
+                    continue;
+
+                var directoryFinalPath = $"{ANDROID_PLUGINS_FOLDER}{directory.Name}" +
+                                         $"{(addAndroidLibraryName ? ANDROID_ADDITIONAL_MANIFEST_FOLDER_END_NAME : String.Empty)}";
+
+                if ( Directory.Exists( directoryFinalPath ) )
+                {
+                    FileUtil.DeleteFileOrDirectory( directoryFinalPath );
+                    FileUtil.DeleteFileOrDirectory( directoryFinalPath + META_FILES );
+                }
+            }
+        }
+
         static void HUFAndroidBuildAssetsResolver()
         {
-            /*var mainManifest = AssetDatabase.LoadAssetAtPath<TextAsset>( ANDROID_PLUGINS_FOLDER + MANIFEST_FULL_NAME );
-
-            if ( mainManifest == null )
-            {
-                HLog.LogError( logPrefix, $"You are missing AndroidManifest.xml in: {PLUGIN_FOLDER}" );
-            }*/
-
             var paths = GetHUFManifests( MANIFEST_COPY_SEARCH );
 
             foreach ( string sourcePath in paths )
@@ -66,34 +137,46 @@ namespace HUF.Utils.Editor.AssetsBuilder
                 if ( !File.Exists( $"{packageAndroidFolder}/{PROJECT_PROPERTY_FULL_NAME}" ) )
                     CreateProjectPropertyFile( packageEndPath );
                 else
-                    FileUtil.CopyFileOrDirectory( $"{packageAndroidFolder}/{PROJECT_PROPERTY_FULL_NAME}",
+                    CopyFiles( $"{packageAndroidFolder}/{PROJECT_PROPERTY_FULL_NAME}",
                         $"{packageEndPath}/{PROJECT_PROPERTY_FULL_NAME}" );
-                HLog.Log( logPrefix, $"Android Manifest Copy from {sourcePath} to {packageEndPath}" );
             }
 
-            paths = GetHUFPackageAndroidAssetsFolder();
+            paths = GetHUFPackageToCopyFolder();
 
-            foreach ( string sourcePath in paths )
+            foreach ( string source in paths )
             {
+                var sourcePath = source + "ToCopy/";
+
+                if ( !Directory.Exists( sourcePath ) )
+                    continue;
+
                 if ( Directory.Exists( sourcePath + "res/" ) )
-                    CopyAndroidResources( sourcePath + "res/", ANDROID_PLUGINS_FOLDER + "res/" );
+                    CopyFiles( sourcePath + "res/", ANDROID_PLUGINS_FOLDER + "res/" );
                 var dir = new DirectoryInfo( sourcePath );
                 var directories = dir.GetDirectories();
-                
+
                 foreach ( var directory in directories )
                 {
                     if ( directory.Name == "res" )
                         continue;
 
-                    var directoryFinalPath = $"{ANDROID_PLUGINS_FOLDER}/{directory.Name}{ANDROID_ADDITIONAL_MANIFEST_FOLDER_END_NAME}";
+                    var directoryFinalPath = $"{ANDROID_PLUGINS_FOLDER}{directory.Name}" +
+                                             $"{( directory.Name != "Firebase" ? ANDROID_ADDITIONAL_MANIFEST_FOLDER_END_NAME : String.Empty )}";
 
                     if ( Directory.Exists( directoryFinalPath ) )
                     {
                         FileUtil.DeleteFileOrDirectory( directoryFinalPath );
                     }
 
-                    FileUtil.CopyFileOrDirectory( $"{sourcePath}{directory.Name}", directoryFinalPath );
+                    CopyFiles( $"{sourcePath}{directory.Name}", directoryFinalPath );
                 }
+            }
+            
+            var firebaseCopyPath = FindSubfolderPath("Assets/Firebase", "ToCopy");
+
+            if ( !firebaseCopyPath.IsNullOrEmpty() )
+            {
+                CopyFiles( firebaseCopyPath, ANDROID_PLUGINS_FOLDER );
             }
         }
 
@@ -105,31 +188,29 @@ namespace HUF.Utils.Editor.AssetsBuilder
                    ANDROID_ADDITIONAL_MANIFEST_FOLDER_END_NAME;
         }
 
-        static void CopyAndroidResources( string source, string destination )
+        static void CopyFiles( string source, string destination )
         {
-            if ( Directory.Exists( destination ) )
+            if ( !Directory.Exists( destination ) )
             {
-                var dir = new DirectoryInfo( source );
-                var files = dir.GetFiles();
-
-                foreach ( var file in files )
-                {
-                    if ( file.Name.Contains( ".meta" ) || File.Exists( $"{destination}/{file.Name}" ) )
-                        continue;
-
-                    FileUtil.CopyFileOrDirectory( $"{source}/{file.Name}", $"{destination}/{file.Name}" );
-                }
-
-                var directories = dir.GetDirectories();
-
-                foreach ( var directory in directories )
-                {
-                    CopyAndroidResources( $"{source}{directory.Name}", $"{destination}{directory.Name}" );
-                }
+                Directory.CreateDirectory( destination );
             }
-            else
+
+            var dir = new DirectoryInfo( source );
+            var files = dir.GetFiles();
+
+            foreach ( var file in files )
             {
-                FileUtil.CopyFileOrDirectory( source, destination );
+                if ( file.Name.Contains( META_FILES ) || File.Exists( $"{destination}/{file.Name}" ) )
+                    continue;
+
+                FileUtil.CopyFileOrDirectory( $"{source}/{file.Name}", $"{destination}/{file.Name}" );
+            }
+
+            var directories = dir.GetDirectories();
+
+            foreach ( var directory in directories )
+            {
+                CopyFiles( $"{source}/{directory.Name}", $"{destination}/{directory.Name}" );
             }
         }
 
@@ -150,11 +231,10 @@ namespace HUF.Utils.Editor.AssetsBuilder
                     s.Contains( packageFolderName + "/" + HUF_ANDROID_ASSET_FOLDER ) );
         }
 
-        static IEnumerable<string> GetHUFPackageAndroidAssetsFolder()
+        static IEnumerable<string> GetHUFPackageToCopyFolder()
         {
-            var dir = new DirectoryInfo( Application.dataPath + "/HUF");
+            var dir = new DirectoryInfo( Application.dataPath + "/HUF" );
             var dirs = dir.GetDirectories();
-            
             List<string> packageAssetsAndroidPath = new List<string>();
 
             foreach ( var directory in dirs )
@@ -163,11 +243,30 @@ namespace HUF.Utils.Editor.AssetsBuilder
 
                 if ( Directory.Exists( path ) )
                 {
-                    packageAssetsAndroidPath.Add( path.Replace( "\\","/" ) +"/" );
+                    packageAssetsAndroidPath.Add( path.Replace( "\\", "/" ) + "/" );
                 }
             }
 
             return packageAssetsAndroidPath;
+        }
+        
+        static string FindSubfolderPath(string parentFolder, string searchFolder)
+        {
+            string[] folders = AssetDatabase.GetSubFolders(parentFolder);
+            string resultFolder = folders.FirstOrDefault(folder => (new DirectoryInfo(folder)).Name == searchFolder);
+            if (string.IsNullOrEmpty(resultFolder) && folders.Length > 0)
+            {
+                string temp;
+                for (int i = 0; i < folders.Length; i++)
+                {
+                    temp = FindSubfolderPath(folders[i], searchFolder);
+                    if (!string.IsNullOrEmpty(temp))
+                    {
+                        return temp;
+                    }
+                }
+            }
+            return resultFolder;
         }
     }
 }
